+++ src/MicroElements.Swashbuckle.FluentValidation/FluentValidationRule.cs
@@ -1,4 +1,7 @@
-﻿using System;
+﻿// Copyright (c) MicroElements. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System;
 using FluentValidation.Validators;
 
 namespace MicroElements.Swashbuckle.FluentValidation
@@ -16,15 +19,15 @@ namespace MicroElements.Swashbuckle.FluentValidation
         /// <summary>
         /// Predicate to match property validator.
         /// </summary>
-        public Func<IPropertyValidator, bool> Matches { get; set; }
+        public Func<IPropertyValidator, bool>? Matches { get; set; }
 
         /// <summary>
         /// Modify Swagger schema action.
         /// </summary>
-        public Action<RuleContext> Apply { get; set; }
+        public Action<RuleContext>? Apply { get; set; }
 
         /// <summary>
-        /// Creates new instance of <see cref="FluentValidationRule"/>.
+        /// Initializes a new instance of the <see cref="FluentValidationRule"/> class.
         /// </summary>
         /// <param name="name">Rule name.</param>
         public FluentValidationRule(string name)